refactor: API 분리
