+++ backend/src/v1/books/books.controller.ts
@@ -8,18 +8,15 @@ import * as errorCode from '~/v1/utils/error/errorCode';
 import ErrorResponse from '~/v1/utils/error/errorResponse';
 import isNullish from '~/v1/utils/isNullish';
 import * as parseCheck from '~/v1/utils/parseCheck';
-import { getAccessToken } from '~/v1/auth/auth.service';
 import * as BooksService from './books.service';
 import * as types from './books.type';
 import LikesService from './likes.service';
 import { searchSchema } from '../users/users.types';
 import { User } from '../DTO/users.model';
 import UsersService from '../users/users.service';
-import { BookListWithSubject, Project } from '../DTO/cursus.model';
 
 const likesService = new LikesService();
 const usersService = new UsersService();
-let accessToken: string;
 
 const pubdateFormatValidator = (pubdate: string | Date) => {
   const regexConditon = /^[0-9]{8}$/;
@@ -464,35 +461,4 @@ export const updateBookDonator = async (
     next(new ErrorResponse(errorCode.FAIL_PATCH_BOOK_BY_UNEXPECTED, status.INTERNAL_SERVER_ERROR));
   }
   return 0;
-};
-
-export const recommandBook = async (
-  req: Request,
-  res: Response,
-  next: NextFunction,
-) => {
-  const { nickname: login } = req.user as any;
-  const limit = Number(req.query.limit);
-  let bookList: BookListWithSubject[] = [];
-  let meta: string[] = [];
-  let userProject: Project[] = [];
-  let userId: string;
-  if (login !== null && login !== undefined) {
-    userId = await BooksService.getIntraId(login);
-    try {
-      userProject = await BooksService.getUserProjectFrom42API(accessToken, userId);
-    } catch (error: any) {
-      if (error.status === 401) {
-        accessToken = await getAccessToken();
-        userProject = await BooksService.getUserProjectFrom42API(accessToken, userId);
-      } else {
-        next(new ErrorResponse(errorCode.UNKNOWN_ERROR, status.INTERNAL_SERVER_ERROR));
-      }
-    }
-    const projectIds: number[] = await BooksService.getRecommendedProject(userProject);
-    const bookIds: number[] = await BooksService.getRecommendedBookIds(projectIds);
-    bookList = await BooksService.getBookListByIds(bookIds, limit);
-    meta = await BooksService.getRecommendMeta();
-  }
-  res.status(status.OK).json({ bookList, meta });
 };
