+++ backend/src/v1/routes/books.routes.ts
@@ -13,114 +13,13 @@ import {
   getLikeInfo,
   updateBookInfo,
   updateBookDonator,
-  recommandBook,
 } from '~/v1/books/books.controller';
 import authValidate from '~/v1/auth/auth.validate';
 import authValidateDefaultNullUser from '~/v1/auth/auth.validateDefaultNullUser';
 import { roleSet } from '~/v1/auth/auth.type';
-import rateLimit from 'express-rate-limit';
 
 export const path = '/books';
 export const router = Router();
-const limiter = rateLimit({
-  windowMs: 60 * 1000, // 1분
-  max: 100, // 1분에 100번
-  message: '너무 많은 요청을 보냈습니다. 잠시 후 다시 시도해주세요.',
-});
-
-router
-  /**
-   * @openapi
-   * /api/books/recommend:
-   *    get:
-   *      summary: 서클 별 추천 도서를 가져온다
-   *      description: 사용자가 속한 서클의 추천 도서를 가져온다
-   *      tags:
-   *      - books
-   *      parameters:
-   *      - name: limit
-   *        in: query
-   *        description: 가져올 추천 도서의 개수
-   *        schema:
-   *          type: integer
-   *        example: 4
-   *      - name: type
-   *        in: query
-   *        description: 이너 서클과 아우터 서클을 구분한다
-   *        schema:
-   *          type: string
-   *          enum: [inner, outer]
-   *      - name: circleOrSubject
-   *        in: query
-   *        description: 서클 또는 과제 명을 받아온다.
-   *        schema:
-   *          type: string
-   *        example: Libft
-   *      responses:
-   *        '200':
-   *          description: 서클 별 추천 도서의 정보를 가져온다.
-   *          content:
-   *            application/json:
-   *              schema:
-   *                type: object
-   *                properties:
-   *                  items:
-   *                    description: 추천 도서들의 목록
-   *                    type: array
-   *                    items:
-   *                      type: object
-   *                      properties:
-   *                        id:
-   *                          description: DB 상의 book_info.id
-   *                          type: integer
-   *                          example: 42
-   *                        title:
-   *                          description: 도서 제목
-   *                          type: string
-   *                          example: "시작하세요! 도커/쿠버네티스"
-   *                        author:
-   *                          description: 저자
-   *                          type: string
-   *                          example: 용찬호
-   *                        publisher:
-   *                          description: 출판사
-   *                          type: string
-   *                          example: 위키북스
-   *                        image:
-   *                          description: 표지 사진
-   *                          type: string
-   *                          example: https://image.kyobobook.co.kr/images/book/xlarge/394/x9791195982394.jpg
-   *                        publishedAt:
-   *                          description: 출판일자
-   *                          type: string
-   *                          format: date
-   *                          example: 2020-01-31
-   *                        subjects:
-   *                          description: 도서와 관련된 과제들
-   *                          type: array
-   *                          example: ["Inception", "Inception-of-Things"]
-   *                  meta:
-   *                    description: 드롭다운에서 선택할 수 있는 서클과 과제 정보
-   *                    type: array
-   *                    example: ["0서클 | Libft", "1서클 | ft_printf", ..., "Outer | ft_ping"]
-   *        '400':
-   *          description: x 서클에서 y 과제의 추천 도서를 가져오려고 했는데, x 서클에 y 과제가 없다.
-   *          content:
-   *            application/json:
-   *              schema:
-   *                type: json
-   *                description: error decription
-   *                example: { errorCode: 400 }
-   *        '500':
-   *          description: 서버 오류
-   *          content:
-   *           application/json:
-   *              schema:
-   *                type: json
-   *                description: error decription
-   *                example: { errorCode: 500 }
-   */
-  .get('/recommend', limiter, authValidate(roleSet.all), recommandBook);
 
 router
   /**
